Implementação do modal de explicação do processo de adoção. Implementação da busca e avaliação e cálculo de média.

+++ PetSaver.Contracts/Paginas/Response/PetPage/AnuncianteContract.cs
@@ -3,6 +3,6 @@
     public class AnuncianteContract
     {
         public string Nome { get; set; }
-        public int? Avaliacao { get; set; }
+        public double? Avaliacao { get; set; }
     }
 }
